requirement.txt
annotated-types==0.5.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.6.0
attrs==23.1.0
Automat==22.10.0
autopep8==1.6.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2023.7.22
chardet==4.0.0
click==8.1.6
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.5
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.1
django-oauth-toolkit==2.2.0
dnspython==2.4.1
duplicity==0.8.21
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.101.0
fasteners==0.14.1
filelock==3.7.1
flake8==6.0.0
future==0.18.2
greenlet==1.1.2
gyp==0.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httptools==0.6.0
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jwcrypto==1.4.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
orjson==3.9.4
packaging==23.1
pandas==2.0.1
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.2.0
protobuf==3.12.4
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyasn1==0.5.0
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyflakes==3.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.4.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.31
sqlparse==0.4.4
starlette==0.27.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.16.3
wadllib==1.3.6
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.0
